move socket from addTeam mutation
